requirements.txt
# knownledge
fastapi
openai
huggingface
modelscope
sentencepiece
# faiss-gpu
faiss-cpu
FlagEmbedding
BCEmbedding
pypdf
docx2txt
unstructured
pdfminer.six
python-docx
mdtex2html>=1.3.0
latex2mathml>=3.77.0
colorama
modelscope
rapidocr_onnxruntime
PyMuPDF
pptx

# langchain
langchain>=0.1.9
langchainhub>=0.1.14
langchain
langchain-community
langchain-core
langchain-experimental
langchain-openai

# ui
streamlit
streamlit-pills
streamlit_extras
streamlit-option-menu
streamlit_nested_layout
streamlit-antd-components

openai
transformers==4.38.1